% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_valid_property}
\alias{is_valid_property}
\title{Is an property valid?}
\usage{
is_valid_property(property)
}
\arguments{
\item{property}{property to compare}
}
\description{
Compates \code{property} against a list of known, valid "graph" properties
and returns \code{TRUE} if valid.
}
