% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{og_parse}
\alias{og_parse}
\title{Parse Open Graph Metadata}
\usage{
og_parse(url)
}
\arguments{
\item{url}{A string representing the URL of the webpage to parse.}
}
\value{
A named vector of Open Graph metadata.
}
\description{
Parse Open Graph Metadata
}
\examples{
og_parse('https://www.rstudio.com')
}
