% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{og_property}
\alias{og_property}
\title{Extract a specific Open Graph property from a webpage.}
\usage{
og_property(url, property)
}
\arguments{
\item{url}{A string representing the URL of the webpage to parse.}

\item{property}{A string indicating the Open Graph property (e.g., "og:title" or "title").}
}
\value{
A string containing the value of the specified property, or NA if not found.
}
\description{
Extract a specific Open Graph property from a webpage.
}
\examples{
og_property('https://www.rstudio.com', 'og:title')
}
