% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{og_schema}
\alias{og_schema}
\title{Open Graph Schema}
\format{
A list of Open Graph schema.
}
\usage{
og_schema
}
\description{
As of 2024-12-05. Based on \url{https://ogp.me/}
}
\examples{
og_schema
}
\keyword{datasets}
