% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{og_validate}
\alias{og_validate}
\title{Validate Open Graph Metadata}
\usage{
og_validate(metadata)
}
\arguments{
\item{metadata}{A named list of Open Graph metadata.}
}
\value{
A logical value indicating whether all required metadata is present.
}
\description{
Check if all required Open Graph properties are present.
}
\examples{
og_parse('https://www.rstudio.com') |>
  og_validate()
}
