% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opengraph-package.R
\docType{package}
\name{opengraph}
\alias{opengraph}
\alias{opengraph-package}
\title{Tools to Mine 'Open Graph'-like Tags From 'HTML' Content}
\description{
Many services such as 'Facebook' and 'Twitter' rely on 'Open Graph'-like tags
(\url{http://ogp.me/}) to build a formalized knowledge graph from 'HTML' content metadata.
Tools are providfed to extract and analyze this data.
}
\author{
Bob Rudis (bob@rud.is)
}
