% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{read_properties}
\alias{read_properties}
\alias{read_properties.character}
\alias{read_properties.html_document}
\alias{read_properties.xml_document}
\alias{read_properties.response}
\alias{read_properties.raw}
\title{Read Open Graph tags from an HTML document}
\usage{
read_properties(x, ...)

\method{read_properties}{character}(x, ...)

\method{read_properties}{html_document}(x, ...)

\method{read_properties}{xml_document}(x, ...)

\method{read_properties}{response}(x, ...)

\method{read_properties}{raw}(x, ...)
}
\arguments{
\item{x}{URL or parsed HTML/XML document}

\item{...}{Additional arguments passed on to methods.}
}
\description{
Read Open Graph tags from an HTML document
}
