% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{og_schema}
\alias{og_schema}
\title{Open Graph Schema}
\usage{
og_schema()
}
\value{
A character vector of Open Graph schema.
}
\description{
As of 2024-12-05. Based on \url{https://ogp.me/}
}
\examples{
og_schema()
}
